requirements.txt
# === Core AI & Gemini ===
google-generativeai>=0.5.2

# === Web Framework (API Backend) ===
fastapi>=0.110.0
uvicorn>=0.29.0

# === Tools / MCP / Utilities ===
mcp>=1.2.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# === Data & JSON Utilities ===
numpy>=1.26.0
pandas>=2.1.0

# === Text Processing & Embeddings ===
sentence-transformers>=2.3.0

# === Notebooks Support ===
jupyter>=1.0.0

# === Testing ===
pytest>=8.0.0
